comments, refactoring to shrink the interface of IncludedService
